管理者ログイン等追加

+++ src/components/pages/Browsing.tsx
@@ -3,15 +3,16 @@ import { Button, Typography } from "@material-ui/core";
 import styles from "./Browsing.module.scss";
 import Header from "../organisms/Header";
 import { useDispatch, useSelector } from "react-redux";
-import { deleteArticle, handleEdit, selectSelectedItems } from "../../features/article/articleSlice";
+import { deletePost, handleEdit, selectSelectedItems } from "../../features/article/articleSlice";
 import { useHistory } from "react-router-dom";
-// import { article } from "../../data/article";
-// import { Article } from "../../types/article";
+import { selectLogin } from "../../features/login/loginSlice";
+
 
 const Browsing = () => {
   const dispatch = useDispatch();
   const history = useHistory()
   const selectedArticle = useSelector(selectSelectedItems);
+  const loginValue = useSelector(selectLogin);
 
   // 編集画面に移動
   const goWriting = () => {
@@ -21,21 +22,9 @@ const Browsing = () => {
 
   // 記事の削除と選択中記事のリセット
   const deleteItem = () => {
-    dispatch(deleteArticle(selectedArticle.id))
-    // dispatch(selectArticle({
-    //   id: "",
-    //   img: "",
-    //   title: "",
-    //   text: "",
-    //   createdAt: "",
-    //   updateTime: "",
-    //   good: 0,
-    // }))
+    dispatch(deletePost(selectedArticle.id))
     history.push("/");
   }
-  const goHome = () => {
-    history.push("/");
-  };
   return (
     <div className={styles.root}>
       <Header />
@@ -45,14 +34,13 @@ const Browsing = () => {
           <Typography variant="subtitle1">{selectedArticle.text}</Typography>
         </div>
         <div className={styles.action}>
-          <Button variant="contained" color="primary" onClick={goWriting}>
+          <Button variant="contained" color="primary" onClick={goWriting} disabled={loginValue.isAdmin === false}>
             編集
           </Button>
-          <Button variant="contained" color="primary" onClick={deleteItem}>
+          <Button variant="contained" color="primary" onClick={deleteItem} disabled={loginValue.isAdmin === false}>
             削除
           </Button>
         </div>
-        <button onClick={goHome}>ホームへ</button>
       </div>
     </div>
   );
